requirements.txt
aniso8601==7.0.0
appdirs==1.4.3
CacheControl==0.12.6
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
future==0.18.2
google-auth==1.14.3
google-auth-oauthlib==0.4.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
gspread==3.6.0
gspread-pandas==2.2.3
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
numpy==1.18.4
oauthlib==3.1.0
packaging==20.3
pandas==1.0.3
pep517==0.8.2
progress==1.5
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.17.1
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
Rx==1.6.1
six==1.14.0
urllib3==1.25.8
webencodings==0.5.1
